parser: parse qualified identifiers The following improvements have been made in the parser: - Types can now have qualified type names - ast.ModuleName has been removed in favour of ast.SelectorExpr for   the sake of consistency across all selectors. - AST types have been split across several files by their category. - Modules and Imports now use ast.SelectorExpr for the module path. - Bugfix: It was not possible to have a BasicType with more than   one argument because it used parseType recursively.
